+++ NEWS.md
@@ -1,3 +1,5 @@
+# keras3 0.1.0
+
 # keras (development version)
 
 # keras 2.13.0